+++ plugins/aves_video_mpv/lib/aves_video_mpv.dart
@@ -1,3 +1,0 @@
-export 'src/controller.dart';
-export 'src/factory.dart';
-export 'src/metadata.dart';